requirements.txt
annotated-types==0.6.0
anyio==3.7.1
azure-common==1.1.28
azure-core==1.29.6
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.105.0
fastapi-azure-auth==4.3.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
isodate==0.6.1
motor==3.3.2
msal==1.26.0
msal-extensions==1.1.0
packaging==23.2
passlib==1.7.4
portalocker==2.8.2
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
